+++ micro-metadb/service/host.go
@@ -463,7 +463,7 @@ func (handler *DBServiceHandler) AllocResources(ctx context.Context, in *dbpb.DB
 		copyResultToRsp(&r, &hostResource)
 		out.Results = append(out.Results, &hostResource)
 	}
-	out.Rs.Code = common.TIEM_SUCCESS
+	out.Rs.Code = int32(common.TIEM_SUCCESS)
 	return nil
 }
 
@@ -491,7 +491,7 @@ func (handler *DBServiceHandler) AllocResourcesInBatch(ctx context.Context, in *
 	for _, result := range resources.BatchResults {
 		var rsp dbpb.DBAllocResponse
 		rsp.Rs = new(dbpb.DBAllocResponseStatus)
-		rsp.Rs.Code = common.TIEM_SUCCESS
+		rsp.Rs.Code = int32(common.TIEM_SUCCESS)
 		for _, r := range result.Results {
 			var hostResource dbpb.DBHostResource
 			copyResultToRsp(&r, &hostResource)
@@ -499,7 +499,7 @@ func (handler *DBServiceHandler) AllocResourcesInBatch(ctx context.Context, in *
 		}
 		out.BatchResults = append(out.BatchResults, &rsp)
 	}
-	out.Rs.Code = common.TIEM_SUCCESS
+	out.Rs.Code = int32(common.TIEM_SUCCESS)
 	return nil
 }
 
@@ -524,6 +524,6 @@ func (handler *DBServiceHandler) RecycleResources(ctx context.Context, in *dbpb.
 		// return nil to use rsp
 		return nil
 	}
-	out.Rs.Code = common.TIEM_SUCCESS
+	out.Rs.Code = int32(common.TIEM_SUCCESS)
 	return nil
 }
